Update README, change license to BSD-3-Clause
Re-licensed to BSD-3-Clause, text as found on
<https://spdx.org/licenses/BSD-3-Clause.html>. This is in expectation
of borrowing a lot of ideas from chibi-scheme, which is also licensed
under a 3-clause BSD license.

A note for the legally-minded: The license text of chibi-scheme is not
100% identical, but the difference is basically just a rewording of
"author" (chibi-scheme) to "copyright holder or contributor" (SPDX
license text). Also note that at least up to and including this
commit, the poco-scheme codebase is not derived of chibi-scheme in any
way.